feat: add send video (#15)
* feat: add send rest

* feat: add view_once when sending video

* feat: upgrade body limit

* feat: change statics folder

* feat: remove copyright

* feat: add thumbnails in send video

* fix: deleting file

* feat: add params compress video

* chore: update openapi

* chore: update readme

* chore: update readme for installation

* chore: update docker and document

* feat: fix validation and update dashboard page

* fix: docker config & utils code

* chore: update docs

* chore: update docs

+++ src/utils/general.go
@@ -13,8 +13,12 @@ func RemoveFile(delaySecond int, paths ...string) error {
 	}
 
 	for _, path := range paths {
-		err := os.Remove(path)
-		return err
+		if path != "" {
+			err := os.Remove(path)
+			if err != nil {
+				return err
+			}
+		}
 	}
 	return nil
 }
@@ -24,7 +28,9 @@ func CreateFolder(folderPath ...string) error {
 	for _, folder := range folderPath {
 		newFolder := filepath.Join(".", folder)
 		err := os.MkdirAll(newFolder, os.ModePerm)
-		return err
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
